Call websocket events instead of emitting
Calling wait's for a response

Without it, they mostly arrive unordered, which isn't a big deal but more importantly, the websocket connection can disconnect prior to all the updates being sent.

This probably slows it down quite a bit, but it should only be noticeable for small RNA sequences that are fast to process, since the websocket event takes up a larger portion of each step. I think this is probably not a bad thing as running small RNA sequences with many steps are less likely now to overload the celery/gunicorn server, but if you want to change this, you could instead continue using emit, and instead at the end of the task, wait for the server to disconnect the client, which you would have to make it only do once it had received all updates.